+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- [[#108](https://github.com/IronCoreLabs/ironoxide/pull/107)]
+  - Fixes bug to allow decryption of 0 and 1 byte documents
+
 ## 0.17.0
 
 - [[#107](https://github.com/IronCoreLabs/ironoxide/pull/107)]